Updated CompressTransformTracks, should fix some incorrect bone rotations on export.
I think some rotations got a flipped bit on the "W" due to a the fact that Math.Sign() can return {-1,0,1}, and because casting from float to int truncates tranform.Rw to a 0, but calculateW is 1, so this line will flip the bits but it looks like the intended logic is for the bits to flip when the signs are opposite.

+++ SSBHLib/Tools/SSBHAnimTrackEncoder.cs
@@ -602,7 +602,7 @@ namespace SSBHLib.Tools
                         ry.DecompressedValue(transform.Ry) * ry.DecompressedValue(transform.Ry) +
                         rz.DecompressedValue(transform.Rz) * rz.DecompressedValue(transform.Rz))));
 
-                    writer.WriteBits(Math.Sign((int)transform.Rw) != Math.Sign((int)calculateW) ? 1 : 0, 1);
+                    writer.WriteBits(Math.Sign(transform.Rw) != Math.Sign(calculateW) ? 1 : 0, 1);
                 }
             }
